// src/models/TeamData.ts
export interface TeamData {
  date_played: string;
  team_score: number;
  team_wickets: number;
  opponent_score: number;
  opponent_wickets: number;
  opponent_team: string;
  overs_bowled: number;
  overs_batted: number;
  ground: string;
  home_game: boolean;
  result: string;
  team: string;
  toss: string;
  outright: string;
  outright_result: string;
}